Rename generateSchedule and generateScheduleButton directives
generateScheduleButton was extracted from generateSchedule. The
latter only contains scheduling options and saved schedules, so
rename to preferences. Then rename generateSchedulesButton to the
less clunky generateSchedules.
