+++ .gitignore
@@ -17,7 +17,7 @@ client/js/angular/courses.service.js
 client/js/angular/confirmEventDeleteDialog.controller.js
 client/js/angular/event.service.js
 client/js/angular/eventDisplay.directive.js
-client/js/angular/generateSchedulesButton.directive.js
+client/js/angular/generateSchedules.directive.js
 client/js/angular/meetingEditor.directive.js
 client/js/angular/reverse.filter.js
 client/js/angular/timePicker.directive.js
