--- assets/static/html/generate_schedules_button.partial.html
@@ -1,28 +1,0 @@
-<!--<div>-->
-  <span ng-if="gSBCtrl.scheduleGenerationStatus.status === 'stale'">
-    <md-button class="md-accent md-raised" ng-click="gSBCtrl.generateAndViewSchedules()">
-      Generate Schedules
-    </md-button>
-  </span>
-  <span ng-if="
-      gSBCtrl.scheduleGenerationStatus.status === 'generating' ||
-      gSBCtrl.scheduleGenerationStatus.status === 'filteringAndReordering'
-    ">
-    <md-button class="md-accent md-raised" ng-click="gSBCtrl.generateAndViewSchedules()">
-      Generating
-    </md-button>
-  </span>
-  <span ng-if="gSBCtrl.scheduleGenerationStatus.status === 'done'">
-    <md-button class="md-accent md-raised" ng-click="gSBCtrl.viewSchedules()">
-      View Schedules
-    </md-button>
-  </span>
-  <span ng-if="gSBCtrl.scheduleGenerationStatus.status === 'failed'">
-    <md-button ng-click="gSBCtrl.viewSchedules()">
-      No Valid Schedules
-      <md-tooltip md-direction="bottom">
-        Make changes to your courses/filtering options to fix this.
-      </md-tooltip>
-    </md-button>
-  </span>
-<!--</div>-->
