+++ assets/static/html/schedule.html
@@ -132,13 +132,13 @@
   </div>
   <div flex></div>
   <div layout="row" layout-align="center" flex-offset="100" layout-padding>
-    <bs-generate-schedules-button></bs-generate-schedules-button>
+    <bs-generate-schedules></bs-generate-schedules>
   </div>
 </div>
 
 <md-content class="right-pane-container" layout="column" flex>
   <md-toolbar>
-    <bs-generate-schedules md-whiteframe="2"></bs-generate-schedules>
+    <bs-preferences md-whiteframe="2"></bs-preferences>
   </md-toolbar>
   <div class="right-pane" id="right-pane" flex layout="column">
     <ui-view>
